Move `objectToAst`
